refactor: interface for order crud
