+++ src/makeline-service/mongodb.go
@@ -4,33 +4,37 @@ import (
 	"context"
 	"crypto/tls"
 	"log"
-	"net/http"
 	"os"
 
+	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
-func connectToMongoDB() (*mongo.Collection, error) {
+type MongoDBOrderRepo struct {
+	db *mongo.Collection
+}
+
+func NewMongoDBOrderRepo() *MongoDBOrderRepo {
 	// Get database uri from environment variable
 	mongoUri := os.Getenv("ORDER_DB_URI")
 	if mongoUri == "" {
 		log.Printf("ORDER_DB_URI is not set")
-		return nil, http.ErrAbortHandler
+		return nil
 	}
 
 	// get database name from environment variable
 	mongoDb := os.Getenv("ORDER_DB_NAME")
 	if mongoDb == "" {
 		log.Printf("ORDER_DB_NAME is not set")
-		return nil, http.ErrAbortHandler
+		return nil
 	}
 
 	// get database collection name from environment variable
 	mongoCollection := os.Getenv("ORDER_DB_COLLECTION_NAME")
 	if mongoCollection == "" {
 		log.Printf("ORDER_DB_COLLECTION_NAME is not set")
-		return nil, http.ErrAbortHandler
+		return nil
 	}
 
 	// get database username from environment variable
@@ -59,7 +63,7 @@ func connectToMongoDB() (*mongo.Collection, error) {
 	mongoClient, err := mongo.Connect(ctx, clientOptions)
 	if err != nil {
 		log.Printf("failed to connect to mongodb: %s", err)
-		return nil, err
+		return nil
 	}
 
 	err = mongoClient.Ping(ctx, nil)
@@ -71,6 +75,96 @@ func connectToMongoDB() (*mongo.Collection, error) {
 
 	// get a handle for the collection
 	collection := mongoClient.Database(mongoDb).Collection(mongoCollection)
+	//defer collection.Database().Client().Disconnect(context.Background())
 
-	return collection, nil
+	return &MongoDBOrderRepo{collection}
+}
+
+func (r *MongoDBOrderRepo) GetPendingOrders() ([]Order, error) {
+	ctx := context.TODO()
+
+	var orders []Order
+	cursor, err := r.db.Find(ctx, bson.M{"status": Pending})
+	if err != nil {
+		log.Printf("Failed to find records: %s", err)
+		return nil, err
+	}
+	defer cursor.Close(ctx)
+
+	// Check if there was an error during iteration
+	if err := cursor.Err(); err != nil {
+		log.Printf("Failed to find records: %s", err)
+		return nil, err
+	}
+
+	// Iterate over the cursor and decode each document
+	for cursor.Next(ctx) {
+		var pendingOrder Order
+		if err := cursor.Decode(&pendingOrder); err != nil {
+			log.Printf("Failed to decode order: %s", err)
+			return nil, err
+		}
+		orders = append(orders, pendingOrder)
+	}
+
+	return orders, nil
+}
+
+func (r *MongoDBOrderRepo) GetOrder(id string) (Order, error) {
+	var ctx = context.TODO()
+
+	singleResult := r.db.FindOne(ctx, bson.M{"orderid": id})
+	var order Order
+	err := singleResult.Decode(&order)
+	if err != nil {
+		log.Printf("Failed to decode order: %s", err)
+		return order, err
+	}
+
+	return order, nil
+}
+
+func (r *MongoDBOrderRepo) InsertOrders(orders []Order) error {
+	ctx := context.TODO()
+
+	var ordersInterface []interface{}
+	for _, o := range orders {
+		ordersInterface = append(ordersInterface, interface{}(o))
+	}
+
+	if len(ordersInterface) == 0 {
+		log.Printf("No orders to insert into database")
+	} else {
+		// Insert orders
+		insertResult, err := r.db.InsertMany(ctx, ordersInterface)
+		if err != nil {
+			log.Printf("Failed to insert order: %s", err)
+			return err
+		}
+
+		log.Printf("Inserted %v documents into database\n", len(insertResult.InsertedIDs))
+	}
+	return nil
+}
+
+func (r *MongoDBOrderRepo) UpdateOrder(order Order) error {
+	var ctx = context.TODO()
+
+	log.Printf("Updating order: %v", order)
+
+	// Update the order
+	updateResult, err := r.db.UpdateMany(
+		ctx,
+		bson.M{"orderid": order.OrderID},
+		bson.D{
+			{Key: "$set", Value: bson.D{{Key: "status", Value: order.Status}}},
+		},
+	)
+	if err != nil {
+		log.Printf("Failed to update order: %s", err)
+		return err
+	}
+
+	log.Printf("Matched %v documents and updated %v documents.\n", updateResult.MatchedCount, updateResult.ModifiedCount)
+	return nil
 }
